requirements.txt
# Django 핵심
Django==5.0.0

# Django REST Framework
djangorestframework==3.14.0

# JWT 인증
djangorestframework-simplejwt==5.3.1

# MySQL 데이터베이스 드라이버
pymysql==1.1.0

# CORS 지원
django-cors-headers==4.3.1

# 환경변수 관리
python-dotenv==1.0.0

# 개발 도구
django-extensions==3.2.3

